Fix SA1402 violation in DrawingsLabConverters.cs
